+++ app/src/main/java/org/android/go/sopt/ui/main/MainActivity.kt
@@ -1,7 +1,8 @@
-package org.android.go.sopt
+package org.android.go.sopt.ui.main
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import org.android.go.sopt.Week1Application
 import org.android.go.sopt.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
@@ -12,8 +13,8 @@ class MainActivity : AppCompatActivity() {
         binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
-        val name = MyApplication.prefs.getString("name", "nothing")
-        val hobby = MyApplication.prefs.getString("hobby", "nothing")
+        val name = Week1Application.prefs.getString("name", "nothing")
+        val hobby = Week1Application.prefs.getString("hobby", "nothing")
         binding.tvName.text = "이름: $name"
         binding.tvHobby.text = "특기: $hobby"
     }
